moex_vsa_backtester/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moex-vsa-backtester"
version = "0.1.0"
description = "VSA Backtesting Engine for MOEX"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "SQLAlchemy>=2.0.0",
    "PyMySQL>=1.1.0",
    "python-dotenv>=1.0.0",
    "ta>=0.10.0",
    "pytest>=7.4.0",
    "loguru>=0.7.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
vsa-backtest = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.setuptools]
packages = ["config", "db", "core", "backtest"]